fix: resolve 1537 pre-existing ruff violations across src/, tests/, and shipped templates
make lint (ruff) was red on main whenever run with a properly-synced venv;
CI's broken swallow-bug (fixed in the previous commit) meant this never
blocked a merge. Categories fixed:

- Safe/unsafe ruff --fix (1051): Optional[X] -> X | None, Dict/List ->
  dict/list, f-strings, import sorting, unused noqa removal, and other
  mechanical modernizations.
- PLW1510 (311): explicit subprocess.run(..., check=False) -- False is
  already the default, so this is a pure no-op formatting fix.
- BLE001 + S110 (94): documented # noqa on the codebase's existing
  defensive fallback/resilience pattern (hooks and orchestrator internals
  that must degrade gracefully rather than crash or propagate).
- DTZ005 (36): datetime.now(UTC) in state/log/backup timestamps, matching
  the project's established UTC convention. Verified no format-sensitive
  test/consumer depends on the naive-local string.
- EXE001 (19): removed vestigial shebangs from files that are always
  invoked via an explicit interpreter or import (never executed directly);
  __init__.py's shebang is load-bearing (PEP 723 inline-script block) and
  was chmod +x'd instead.
- SIM102 + SIM117 (25): collapsed nested if/with blocks with no behavior
  change.
- F401, RUF012, B017, PLW0406, PLC0206, PERF402 (17): removed dead
  imports, added ClassVar to mutable class-attribute defaults, narrowed
  two overly-broad pytest.raises(Exception) to the actual raised type,
  documented a deliberate self-import, and simplified two loops.
- B008 and TRY004 (56): disabled via pyproject.toml config rather than
  restructuring code -- typer.Option/Argument-as-default is the framework's
  required API, and this codebase deliberately raises ValueError (not
  TypeError) from its input validators, a tested/documented contract that
  TRY004 disagreed with.

Also restored 6 `# pyright: ignore[reportMissingImports]` suppressions
that the RUF100 unused-noqa cleanup collaterally stripped (or, for
multi-line imports, left on the wrong line) when it removed an adjacent
`# noqa: E402`/`# noqa: PLC0415` sharing the same comment.

All fixes verified behavior-preserving: full pytest suite green
(4094 passed) before and after each category, make check clean end to
end (ruff, mypy, pyright, hook-lint, tests, render-parity).

+++ src/mapify_cli/token_budget.py
@@ -220,7 +220,7 @@ def format_compact_instruction(used: int, threshold: int, focus: str) -> str:
     ``ralph-context-pruner`` and ``post-compact-context``, so the assistant
     can recognise where the message came from.
     """
-    pct = int(round(100 * used / threshold)) if threshold > 0 else 0
+    pct = round(100 * used / threshold) if threshold > 0 else 0
     # Fallback must match the documented default in
     # ``docs/context-compression-plan.md`` (Defaults table) so the user gets
     # the same /compact instruction whether they leave compression_focus blank
